Refactor as component
